mailserver: basic services
First commit to add basic services - dovecot and postfix.
Also included is a minimal caddy webserver for certificate request
and renewal.

Some helper scripts are added to add and remove users from the mail
server as well as to update aliases.

Initial set-up is automated via environment file.

A README is provided to get people going.

Signed-off-by: Thilo Fromm <thilo.alexander@gmail.com>