mailserver-monitoring-network.service: fix race with mailserver network
mailserver-monitoring-network is a one-shot service which sets up internal docker networking for all monitoring services. All actual metrics services depend on this network (service).

There was a race condition with "systemctl restart mailserver" which had this service start in parallel with the umbrella "mailserver-monitoring" and therefore in parallel with the shutdown and restart of the main mailserver service. This could lead to the monitoring networking connecting to the soon-to-be-shut-down mailserver network instead of the "new" mailserver network after restart, leading to metrics not being scraped and Grafana becoming unreachable.