contrib/mailman.yaml: make logwatcher use mailman-network
The mailman contrib involuntarily created a separate network "mailman_default" for the logwatcher container as this container was not explicitly assigned to the external mailserver-network.

While this had no serious implications, starting mailman caused a default network to unnecessarily be created for logwatcher. This change assigns logwatcher to mailman-network to prevent this.